Don't implement Deref for unit types.

+++ src/ext4/inode/inode.rs
@@ -75,7 +75,10 @@ impl Inode {
             .contains(ReadOnlyCompatibleFeatures::METADATA_CHECKSUMS)
         {
             Some(ext4_crc32c(
-                ext4_crc32c(superblock.checksum_seed(), inode_number.to_le_bytes()),
+                ext4_crc32c(
+                    superblock.checksum_seed(),
+                    u32::from(inode_number).to_le_bytes(),
+                ),
                 &buffer.as_ref()[layout::generation::OFFSET..][..layout::generation::SIZE.unwrap()],
             ))
         } else {
